first draft dyn inventory support
